Updated
Worked on extracting BT (Behavior Tree) nodes into separate Behavior Trees to make the BT cleaner.

Working on attacking code. Appears to be an issue with the attacks not getting reset at a certain time. Need to look more into that.

Temp Fix: Block.  It will clear the boolean not being reset.

Which is why I need to turn to enums...